Combining the tests
